Made the current system use the new packet system

+++ CrazyCanvas/Include/Events/MatchEvents.h
@@ -8,7 +8,7 @@
 struct OnFlagDeliveredEvent : public LambdaEngine::Event
 {
 public:
-	inline OnFlagDeliveredEvent(uint32 teamIndex) : 
+	inline OnFlagDeliveredEvent(uint8 teamIndex) : 
 		Event(),
 		TeamIndex(teamIndex)
 	{
